Template for custom strategies and creation of custom players
You can now add a custom strategy according to "template.py" in
"custom-strategies". In "gui/util.py" there is a function "createPlayer"
that takes the file path to a strategy along with custom players name
and returns an axelrod.Player object that represents a player with that
strategy. This object can be used like any other Player object in
axelrod's library.